requirements.txt
colorama==0.4.6
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pytest==7.3.2
tomli==2.0.1
numpy~=1.24.3